Move REDCapEmailList to redcap-api
