+++ requirements.txt
@@ -26,6 +26,6 @@ types-PyYAML>=6.0.12
 urllib3 >= 2.2.1
 ratelimit>=2.2.1
 ratelimit-types>=0.0.1
-redcap_api@ https://github.com/naccdata/redcap-api/releases/download/v0.1.1/redcap_api-0.1.1-py3-none-any.whl
+redcap_api@ https://github.com/naccdata/redcap-api/releases/download/v0.2.0/redcap_api-0.2.0-py3-none-any.whl
 nacc_form_validator@ https://github.com/naccdata/nacc-form-validator/releases/download/v0.5.2/nacc_form_validator-0.5.2-py3-none-any.whl
 nacc_attribute_deriver@ https://github.com/naccdata/nacc-attribute-deriver/releases/download/v1.4.0/nacc_attribute_deriver-1.4.0-py3-none-any.whl
